Prevent reinstallation of an already in-use group key
Track the current GTK and IGTK that is in use and when receiving a
(possibly retransmitted) Group Message 1 or WNM-Sleep Mode Response, do
not install the given key if it is already in use. This prevents an
attacker from trying to trick the client into resetting or lowering the
sequence counter associated to the group key.

Signed-off-by: Mathy Vanhoef <Mathy.Vanhoef@cs.kuleuven.be>

Git-commit: cb5132bb35698cc0c743e34fe0e845dfc4c3e410
Git-repo: git://w1.fi/srv/git/hostap.git
Change-Id: I42fb6bab926cec0ab51bc7f9856332a318ca0a67
CRs-Fixed: 2128039
Conflicts:
	src/common/wpa_common.h
	src/rsn_supp/wpa.c
[Backport : Add WPA_IGTK_MAX_LEN definition and
 Refine wpa_supplicant_install_igtk per current impl]