Wifi: Add debug enhancement for hostapd
The default log level of hostapd is MSG_INFO. We often need more logs to
debug some problems. In this change, we add a API of set_log_level.
Everytime launching hostapd, we will check the system property of
"vendor.qcom.wifi.debug", if the value is 1, we will set log level to
MSG_EXCESSIVE, otherwise，it will keep the default level MSG_INFO.

CRs-Fixed: 2173654
Change-Id: I3fbcc282fafc12f8df6903b4a4fdae363cd1db97